vendor update, no tests
This once more updates to a recent version of the e2e-refactor-pr, now
based on Kubernetes 1.12. This time, test/e2e/storage does not get
imported, because we want to demonstrate that no cloud-provider
specific code is needed by the framework itself. This allows us to get
rid of github.com/aws.