use modified `testsuites` with less dependencies
PR https://github.com/kubernetes/kubernetes/pull/70862 reduces the
number of packages that `testsuites` depends on. This allows us to get
rid of `storage/vsphere` and the tests defined by it.

It also introduces a simpler configuration mechanism. We use that to
pin the hostpath pods to a single node.